Minor gui fixes. Separate files for GUI and control
